Правка замечаний по предыдущим заданиям
Выполнены правки по заданиям 1 и 4
Добавлена ссылка на профиль куратора (задание 1)
Задание 4: Файлы переименованы согласно критериям оценивания проектов.
Исправлены замечания от eslint (одинарные кавычки и отступы)
