+++ Readme.md
@@ -1,7 +1,7 @@
 # Личный проект «Что посмотреть»
 
 - Студент: [Александр Согрин](https://up.htmlacademy.ru/univer-js3-urfu/1/user/1905911).
-- Наставник: `Юлия Корецкая`.
+- Наставник: [Юлия Корецкая](https://htmlacademy.ru/profile/id11886).
 
 ---
