+++ project/src/components/FilmCard/film-card.tsx
