+++ project/src/components/app/app.tsx
@@ -1,5 +1,5 @@
-import MainPage from "../../pages/MainPage";
-import MainPageProps from "./../../types/MainPageProps";
+import MainPage from '../../pages/MainPage/main-page';
+import MainPageProps from '../../types/main-page-props';
 
 function App({ name, genre, released }: MainPageProps): JSX.Element {
   return <MainPage name={name} genre={genre} released={released} />;
