+++ project/src/index.tsx
@@ -1,14 +1,14 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./components/app/app";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './components/app/app';
 
 const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLElement
+  document.getElementById('root') as HTMLElement
 );
 
 const testProps = {
-  name: "The Grand Budapest Hotel",
-  genre: "Comedy",
+  name: 'The Grand Budapest Hotel',
+  genre: 'Comedy',
   released: 2014,
 };
