+++ project/src/types/main-page-props.ts
